Make sure coefficients always add up to 1.0
When we shrink an image and scale the filter, we sometimes end up
with 6/10/14/etc. taps, which are not multiples of four, and don't
end up giving a total weight of 1.0.

In this case, we calculate how much of an error there is, and we
fudge the center sample weight to make sure that all weights add
up to 1.0.

This approach is also used in the zoom image resizer.